TinyMCE per Block: Adding the align images controls
