+++ tinymce-per-block/src/assets/stylesheets/main.scss
@@ -2,7 +2,6 @@
 @import 'animations';
 @import '~blocks/text-block/style';
 @import '~blocks/image-block/style';
-@import '~blocks/image-caption-block/style';
 @import '~blocks/heading-block/style';
 @import '~blocks/paragraph-block/style';
 @import '~blocks/quote-block/style';
