+++ tinymce-per-block/src/blocks/image-block/_style.scss
@@ -1,4 +1,50 @@
-.image-block__display {
-	display: block;
-	max-width: 100%;
+.image-caption-block {
+	position: relative;
+
+	&.align-full-width {
+		margin-left: calc(50% - 50vw);
+		width: 100vw;
+		max-width: none;
+		padding-left: 0;
+		padding-right: 0;
+
+		img {
+			width: 100%;
+		}
+	}
+
+	&.align-left {
+		float: left;
+		width: 340px;
+		z-index: 1;
+	}
+
+	&.align-right {
+		float: right;
+		width: 340px;
+		z-index: 1;
+	}
+
+
+	.image-caption-block__display {
+		display: block;
+		max-width: 100%;
+	}
+
+	.image-caption-block__caption textarea,
+	.image-caption-block__caption .textarea-mirror {
+		margin-top: 10px;
+		width: 100%;
+		border: none;
+		font: inherit;
+		font-family: "Merriweather", serif;
+		font-weight: 300;
+		font-size: 14px;
+		color: $gray-dark-300;
+		resize: none;
+
+		&:focus {
+			outline: 0;
+		}
+	}
 }
