+++ tinymce-per-block/src/external/dashicons/icons/image-align-left.js
@@ -1,0 +1,7 @@
+/**
+ * External dependencies
+ */
+import { createElement } from 'wp-elements';
+
+// This is a gridicon
+export default () => <svg className="dashicon" height="24" width="24" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><g><path d="M3 5h18v2H3V5zm0 14h18v-2H3v2zm0-4h8V9H3v6zm10 0h8v-2h-8v2zm0-4h8V9h-8v2z"></path></g></svg>;
