+++ tinymce-per-block/src/external/dashicons/icons/image-align-right.js
@@ -1,0 +1,7 @@
+/**
+ * External dependencies
+ */
+import { createElement } from 'wp-elements';
+
+// This is a gridicon
+export default () => <svg className="dashicon" height="24" width="24" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><g><path d="M21 7H3V5h18v2zm0 10H3v2h18v-2zm0-8h-8v6h8V9zm-10 4H3v2h8v-2zm0-4H3v2h8V9z"></path></g></svg>;
