+++ tinymce-per-block/src/external/dashicons/index.js
@@ -14,3 +14,7 @@ export { default as FormatImageIcon } from './icons/format-image';
 export { default as EditorBoldIcon } from './icons/editor-bold';
 export { default as EditorItalicIcon } from './icons/editor-italic';
 export { default as EditorStrikethroughIcon } from './icons/editor-strikethrough';
+export { default as ImageNoAlignIcon } from './icons/image-align-left';
+export { default as ImageAlignLeftIcon } from './icons/image-align-left';
+export { default as ImageAlignRightIcon } from './icons/image-align-right';
+export { default as ImageFullWidthIcon } from './icons/image-full-width';
